+++ packages/lib/package.json
@@ -19,7 +19,8 @@
     "markdown-it": "^13.0.1",
     "posthog-node": "^3.1.1",
     "server-only": "^0.0.1",
-    "tailwind-merge": "^1.14.0"
+    "tailwind-merge": "^1.14.0",
+    "@paralleldrive/cuid2": "^2.2.1"
   },
   "devDependencies": {
     "@formbricks/tsconfig": "*",
